Refactor Signup page to use SignupForm component
